First commited as before_18281_rev1

+++ lttng/org.eclipse.linuxtools.ctf.core/src/org/eclipse/linuxtools/ctf/core/event/types/FloatDefinition.java
@@ -88,7 +88,7 @@ public class FloatDefinition extends Definition {
     public void read(BitBuffer input) {
         /* Offset the buffer position wrt the current alignment */
         int align = (int) declaration.getAlignment();
-        int pos = input.position() + ((align - (input.position() % align)) % align);
+        long pos = input.position() + ((align - (input.position() % align)) % align);
         input.position(pos);
 
         final int exp = declaration.getExponent();
